Added global exception handler for adding exceptions and created work order entity.

+++ src/main/java/com/cts/dao/TaskDAO.java
@@ -3,7 +3,7 @@ package com.cts.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.cts.model.Task;
 
-public interface TaskDAO extends JpaRepository<Task, Integer> {
+public interface TaskDAO extends JpaRepository<Task, Long> {
 
 }
